adjust Forum component to be hooked up to redux rather than the PostIndex component
