+++ frontend/src/components/forumComponents/posts/postIndexItem.js
@@ -1,6 +1,10 @@
 import React from 'react'
 
 const PostIndexItem = ({post}) => {
+  // logic for determining if the author of the post is the current user
+  // if it is, allow them to delete or edit the post
+    // if(post._user === currentUser) {}
+
   return (
     <div>
       <h3>{post.title}</h3>
