+++ frontend/src/reducers/postsReducer.js
@@ -1,14 +1,20 @@
+import _ from 'lodash';
 import { FETCH_POSTS } from '../actions/postActions';
 
-const postsReducer = (state = [], action) => {
+const postsReducer = (state = {}, action) => {
   Object.freeze(state);
 
   switch(action.type) {
 
     case FETCH_POSTS:
-      const posts = action.payload;
-      console.log(posts);
-      return posts;
+    const posts = action.payload;
+    const postsObject = {};
+
+    for(let post of posts) {
+      postsObject[post._id] = post;
+    }
+
+    return _.merge({}, state, postsObject);
 
     default:
       return state;
